fix(shared-prefs): cast double→float explicitly in setDouble/getDouble
The shared_preferences plugin fix from 4ba0291 (and consequently
1a8d6ba which built on top) failed at Build Release APK with
two Java compile errors:

  SharedPreferencesPlugin.java:87: incompatible types: possible
      lossy conversion from double to float

  SharedPreferencesPlugin.java:125: same

Java's type system requires an explicit cast for narrowing
conversions (double → float). Fixed both sites:
  - setDouble: added (float) cast on the value before putFloat
  - getDouble: changed the 0.0 default to 0.0f (getFloat's second
    argument is float, not double)

Note: this also re-ships the 1a8d6ba NSFW-subreddit commit
which the previous build did not produce an APK for.

+++ android/app/src/main/java/io/flutter/plugins/sharedpreferences/SharedPreferencesPlugin.java
@@ -83,8 +83,9 @@ public class SharedPreferencesPlugin implements FlutterPlugin, MethodChannel.Met
                     break;
                 case "setDouble":
                     // SharedPreferences has no native double — encode as float
-                    // (same trick the upstream plugin uses)
-                    editor.putFloat(prefixed(keyOf(call)), ((Number) call.argument("value")).doubleValue()).apply();
+                    // (same trick the upstream plugin uses). Explicit (float)
+                    // cast required by Java because double→float is narrowing.
+                    editor.putFloat(prefixed(keyOf(call)), (float) ((Number) call.argument("value")).doubleValue()).apply();
                     result.success(true);
                     break;
                 case "setStringList": {
@@ -122,7 +123,9 @@ public class SharedPreferencesPlugin implements FlutterPlugin, MethodChannel.Met
                     break;
                 case "getDouble":
                     // No native double; decode from the float we stored.
-                    result.success((double) prefs.getFloat(prefixed(keyOf(call)), 0.0));
+                    // 0.0f (not 0.0) — Java distinguishes float and double
+                    // literals and getFloat needs a float default.
+                    result.success((double) prefs.getFloat(prefixed(keyOf(call)), 0.0f));
                     break;
                 case "getStringList": {
                     java.util.Set<String> set = prefs.getStringSet(prefixed(keyOf(call)), null);
